grc-tech/requirements.txt
psutil
tabulate
colorama
dnspython
pandas
